Stats for max min average and all
"ave", "max", "min" give back the request stat.
"stats" gives back all 3 statistics
The code keeps counters for all 3 stats, even if only 1 stat is
requested, because they are very easy stats and it would be nearly as
much compute time to do extra if statements to avoid the unneeded parts
of the calculation.