Make the Redraw button always requery the data
I can blank out either the from or to date picker (or both) and the API
happily accepts a query with only 1 (or neither) of from and to
parameters. If "to" is not given then the API returns all the available
data from "from". Similar sensible behaviour if "from" is omitted.
But the data was only being re-queried when a date picker field changed.
Actually we want to also re-query when the Redraw button is pressed.
This is particularly needed when the "to" date picker is empty - we can
press Redraw and the data will be brought "up to current" without having
to touch the date picker controls.