Return lower and upper as the actual first and last timestamps
When we ask for some big date-time range in from and to, it is nice to
know the actual date-time range returned, so we can show this to the
user (e.g. in the mouse over box that describes the Temperature graph).
After this change, the lower and upper attributes are returned as the
actual least and greatest time stamps in the returned data set.
"from" and "to" are also returned, in case something cares about them
and wants to confirm exactly what was asked for.
Var names lower and upper changes to from_timestamp and to_timestamp to
(hopefully) reduce potential confusion.