Return from to lower upper for ave min max
This makes the API consistent. Whenever a statistic or range of data is
requested, the result includes:
from - the starting time that the caller requested
to - the ending time that the caller requested
lower - the actual lowest time in the database within the requested
range
uper - the actual highest time in the database within the requested
range

So any app can use that as it likes to inform the user about the actual
range of the data, as distinct from the requested range.