Handle lower and upper when no matching results
If a request for a temperature array was done for a time range that had
no data, then the existing code would get an array index error trying to
access results[1] and results[-1]
This fixes the problem.
I also confirmed that a query that has just a single sample in the time
period works OK - results[1] and results[-1] are the same thing, which
is good.