Start doing API testing with Travis
These are the basics to:
a) Setup stuff in Travis - .travis.yml
b) Put a config.py in place to use for the test server - config.test.py
c) Start the test server in the background
d) Do tests using Frisby under jasmine-node

This is just the basics. It only tests the api/info endpoint at the
moment. From this point, we can add tests for the rest of the endpoints,
which will need ways to make the server generate some test data rather
than just random stuff.