Add sleep after run_server
Sometimes the tests get a 500 not found/timeout response in the Travis container.
We need to run the server in the background, so it can sit providing the data and API while the actual tests run. But we need to give it a chance to get running and generate the first "random" temperature reading, before trying to access the API.
For the moment this will do the job.
In future we could have the server write some flag file or... after it is up and running properly, then wait on that file in the main script here.