Corrections to min and max selection
In the Travis docker image, the queries that combined the selection of a
Temperature record with the selection of a group statistic (min, max or
count) were causing problems.
filtered.first() was always returning just the last Temperature record
in the set (which happens to be the one with the biggest timestamp),
along with the statistic(s) (count, min and/or max value).
I have separated out the queries so that the queries for an actual
Temperature record are done for just that, with order_by() to get the
min or max record at the front of the list.
Now it works fine on Travis, and I also tried a few queries manually in
my Ubuntu VM and they are working fine.
Actually I don't understand how it was working for us on the RPi, must
have been something special in the way the "mixed" queries happened.
The test spec is enhanced to check all the expected min and max from the
test cases.