Tests for limit parameter (#56)
Enhance the code that calculates the jump factor to:
a) If limit = 1 then this is a special case where we return the middle sample of the range.
b) Calculate up to (limit-1) points from the first (n-1) points, then always include the last point, making a maximum of up to limit points.
Previously it would sometimes return limit+1 points (e.g. in live data with a limit of 300, it sometimes returned 301 points, and with various limits out of the 50-point test data, it would return limit+1 points).
I will add a few more test cases later tonight.