Workaround to display max or min f they are at extremities
of the graph.
If the max or min was exactly at the lowest or highest timestamp, then
the graphic was not displaying the marker.
This code adjusts the timestamp of the marker to be a little point
inside the graph limits, and then the graphics library happily displays
the marker.